Sửa giao diện

+++ Web_BHGD/Repositories/EFProductRepository.cs
@@ -4,45 +4,59 @@ using Microsoft.EntityFrameworkCore;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-public class EFProductRepository : IProductRepository
+namespace Web_BHGD.Repositories
 {
-    private readonly ApplicationDbContext _context;
-
-    public EFProductRepository(ApplicationDbContext context)
+    public class EFProductRepository : IProductRepository
     {
-        _context = context;
-    }
+        private readonly ApplicationDbContext _context;
 
-    public async Task<IEnumerable<Product>> GetAllAsync()
-    {
-        // return await _context.Products.ToListAsync();
-        return await _context.Products
-            .Include(p => p.Category) // Include thông tin về category
-            .ToListAsync();
-    }
+        public EFProductRepository(ApplicationDbContext context)
+        {
+            _context = context;
+        }
 
-    public async Task<Product> GetByIdAsync(int id)
-    {
-        return await _context.Products.Include(p =>
-       p.Category).FirstOrDefaultAsync(p => p.Id == id);
-    }
+        public async Task<IEnumerable<Product>> GetAllAsync()
+        {
+            return await _context.Products
+                .Include(p => p.Category)
+                .ToListAsync();
+        }
 
-    public async Task AddAsync(Product product)
-    {
-        _context.Products.Add(product);
-        await _context.SaveChangesAsync();
-    }
+        public async Task<Product> GetByIdAsync(int id)
+        {
+            return await _context.Products
+                .Include(p => p.Category)
+                .FirstOrDefaultAsync(p => p.Id == id);
+        }
 
-    public async Task UpdateAsync(Product product)
-    {
-        _context.Products.Update(product);
-        await _context.SaveChangesAsync();
-    }
+        public async Task<IEnumerable<Product>> GetByCategoryIdAsync(int categoryId)
+        {
+            return await _context.Products
+                .Include(p => p.Category)
+                .Where(p => p.CategoryId == categoryId)
+                .ToListAsync();
+        }
 
-    public async Task DeleteAsync(int id)
-    {
-        var product = await _context.Products.FindAsync(id);
-        _context.Products.Remove(product);
-        await _context.SaveChangesAsync();
+        public async Task AddAsync(Product product)
+        {
+            _context.Products.Add(product);
+            await _context.SaveChangesAsync();
+        }
+
+        public async Task UpdateAsync(Product product)
+        {
+            _context.Products.Update(product);
+            await _context.SaveChangesAsync();
+        }
+
+        public async Task DeleteAsync(int id)
+        {
+            var product = await _context.Products.FindAsync(id);
+            if (product != null)
+            {
+                _context.Products.Remove(product);
+                await _context.SaveChangesAsync();
+            }
+        }
     }
 }
